// apps/web/src/lib/api/generated/types/groupMemberResponseDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Movies App API
 * REST API for Movies application

**Links:**
- [OpenAPI JSON](/api/docs/json)
- [OpenAPI YAML](/api/docs/yaml)
 * OpenAPI spec version: 1.0.0
 */
import type { GroupMemberResponseDtoRole } from './groupMemberResponseDtoRole';
import type { GroupMemberResponseDtoUser } from './groupMemberResponseDtoUser';

export interface GroupMemberResponseDto {
	id: number;
	groupId: number;
	userId: number;
	role: GroupMemberResponseDtoRole;
	createdAt: string;
	updatedAt: string;
	user?: GroupMemberResponseDtoUser;
}